refactor: TheArticle.vueをorganisms/へ
