+++ src/main.ts
@@ -1,7 +1,7 @@
 import { ViteSSG } from "vite-ssg";
 import generatedRoutes from "virtual:generated-pages";
 import { setupLayouts } from "virtual:generated-layouts";
-import TheArticle from "@/components/TheArticle.vue";
+import TheArticle from "@/components/organisms/TheArticle.vue";
 import DateText from "@/components/atoms/DateText.vue";
 import VueGtag from "vue-gtag";
 import App from "./App.vue";
